// Client/src/types/packets.ts
export enum ClientPacketIds {
  ClientVersion = 0,
  Disconnect = 1,
  KeepAlive = 2,
  NewAccount = 3,
  ChangePassword = 4,
  Login = 5,
  NewCharacter = 6,
  DeleteCharacter = 7,
  StartGame = 8,
  LogOut = 9,
  Turn = 10,
  Walk = 11,
  Run = 12,
  Chat = 13,
  MoveItem = 14,
  StoreItem = 15,
  TakeBackItem = 16,
  MergeItem = 17,
  EquipItem = 18,
  RemoveItem = 19,
  RemoveSlotItem = 20,
  SplitItem = 21,
  UseItem = 22,
  DropItem = 23,
  DepositRefineItem = 24,
  RetrieveRefineItem = 25,
  RefineCancel = 26,
  RefineItem = 27,
  CheckRefine = 28,
  ReplaceWedRing = 29,
  DepositTradeItem = 30,
  RetrieveTradeItem = 31,
  TakeBackHeroItem = 32,
  TransferHeroItem = 33,
  DropGold = 34,
  PickUp = 35,
  RequestMapInfo = 36,
  RequestMonsterInfo = 37,
  RequestNPCInfo = 38,
  RequestItemInfo = 39,
  TeleportToNPC = 40,
  SearchMap = 41,
  Inspect = 42,
  Observe = 43,
  ChangeAMode = 44,
  ChangePMode = 45,
  ChangeTrade = 46,
  Attack = 47,
  RangeAttack = 48,
  Harvest = 49,
  CallNPC = 50,
  BuyItem = 51,
  SellItem = 52,
  CraftItem = 53,
  RepairItem = 54,
  BuyItemBack = 55,
  SRepairItem = 56,
  MagicKey = 57,
  Magic = 58,
  SwitchGroup = 59,
  AddMember = 60,
  DellMember = 61,
  GroupInvite = 62,
  NewHero = 63,
  SetAutoPotValue = 64,
  SetAutoPotItem = 65,
  SetHeroBehaviour = 66,
  ChangeHero = 67,
  TownRevive = 68,
  SpellToggle = 69,
  ConsignItem = 70,
  MarketSearch = 71,
  MarketRefresh = 72,
  MarketPage = 73,
  MarketBuy = 74,
  MarketGetBack = 75,
  MarketSellNow = 76,
  RequestUserName = 77,
  RequestChatItem = 78,
  EditGuildMember = 79,
  EditGuildNotice = 80,
  GuildInvite = 81,
  GuildNameReturn = 82,
  RequestGuildInfo = 83,
  GuildStorageGoldChange = 84,
  GuildStorageItemChange = 85,
  GuildWarReturn = 86,
  MarriageRequest = 87,
  MarriageReply = 88,
  ChangeMarriage = 89,
  DivorceRequest = 90,
  DivorceReply = 91,
  AddMentor = 92,
  MentorReply = 93,
  AllowMentor = 94,
  CancelMentor = 95,
  TradeRequest = 96,
  TradeReply = 97,
  TradeGold = 98,
  TradeConfirm = 99,
  TradeCancel = 100,
  EquipSlotItem = 101,
  FishingCast = 102,
  FishingChangeAutocast = 103,
  AcceptQuest = 104,
  FinishQuest = 105,
  AbandonQuest = 106,
  ShareQuest = 107,
  AcceptReincarnation = 108,
  CancelReincarnation = 109,
  CombineItem = 110,
  AwakeningNeedMaterials = 111,
  AwakeningLockedItem = 112,
  Awakening = 113,
  DisassembleItem = 114,
  DowngradeAwakening = 115,
  ResetAddedItem = 116,
  SendMail = 117,
  ReadMail = 118,
  CollectParcel = 119,
  DeleteMail = 120,
  LockMail = 121,
  MailLockedItem = 122,
  MailCost = 123,
  UpdateIntelligentCreature = 124,
  IntelligentCreaturePickup = 125,
  RequestIntelligentCreatureUpdates = 126,
  AddFriend = 127,
  RemoveFriend = 128,
  RefreshFriends = 129,
  AddMemo = 130,
  GuildBuffUpdate = 131,
  NPCConfirmInput = 132,
  GameshopBuy = 133,
  ReportIssue = 134,
  GetRanking = 135,
  Opendoor = 136,
  GetRentedItems = 137,
  ItemRentalRequest = 138,
  ItemRentalFee = 139,
  ItemRentalPeriod = 140,
  DepositRentalItem = 141,
  RetrieveRentalItem = 142,
  CancelItemRental = 143,
  ItemRentalLockFee = 144,
  ItemRentalLockItem = 145,
  ConfirmItemRental = 146,
  GuildTerritoryPage = 147,
  PurchaseGuildTerritory = 148,
  DeleteItem = 149,
  UnlockStorage = 150,
  SetStoragePassword = 151,
  RemoveStoragePassword = 152,
}